Fix bug when pushing a malformed payload
Before this fix, if you tried to `PUSH` an element using some
of the following payloads:

* `{"elent":8}`
* `{"foo":8}`
* `{1:8}`

The server wouldn't complain when parsing them, and would store a
`null` on top of the Stack. Same behavior when trying to set some config
value.

With this fix, we now check that the payload contains the word `element`
as key. Otherwise, we return an error when decoding it, that will become
a `400 Bad Request` in `pilad`.